+++ src/Server/InverterService/StatusRetriever.cs
@@ -24,7 +24,10 @@ internal class StatusRetriever : BackgroundService
         {
             if (queue.IsAcceptingCommands)
             {
-                cmd.Result.PVMaxCapacity = userSettings.PVMaxCapacity; //feels hacky. find a better solution.
+                //feels hacky. find a better solution.
+                cmd.Result.BatteryCapacity = userSettings.BatteryCapacity;
+                cmd.Result.PVMaxCapacity = userSettings.PVMaxCapacity;
+
                 queue.AddCommands(cmd);
                 _ = db.UpdateTodaysPVGeneration(cmd, c);
             }