+++ src/Server/Persistance/Settings/UserSettings.cs
@@ -6,12 +6,14 @@ public class UserSettings
 {
     public int Id { get; set; } = 1;
     public int PVMaxCapacity { get; set; } = 10;
+    public int BatteryCapacity { get; set; } = 100;
     public int SunlightStartHour { get; set; } = 6;
     public int SunlightEndHour { get; set; } = 18;
 
     public SystemSpec ToSystemSpec() => new()
     {
         PVMaxCapacity = PVMaxCapacity,
+        BatteryCapacity = BatteryCapacity,
         SunlightStartHour = SunlightStartHour,
         SunlightEndHour = SunlightEndHour
     };
@@ -20,6 +22,7 @@ public class UserSettings
     {
         Id = 1;
         PVMaxCapacity = spec.PVMaxCapacity;
+        BatteryCapacity = spec.BatteryCapacity;
         SunlightStartHour = spec.SunlightStartHour;
         SunlightEndHour = spec.SunlightEndHour;
     }