+++ src/changelog.md
@@ -1,4 +1,5 @@
 ## changelog
 - display heatsink temperature
+- progress bar for battery discharge C rate
 - increase grid usage tolerance level to 100w
 - blink temperature text if over 50 celcius